Decomposing Session in progress
